+++ R/check_dup_bp.R
@@ -8,18 +8,53 @@
 #' @importFrom data.table key
 #' @importFrom data.table setkey
 check_dup_bp <- function(sumstats_dt,
+                         bi_allelic_filter,
+                         indels,
                          path,
                          log_folder_ind,
                          check_save_out,
                          tabix_index,
                          nThread,
                          log_files) {
-    BP <- CHR <- NULL
+    BP <- CHR <- A1 <- A2 <- SNP <- NULL
     col_headers <- names(sumstats_dt)
-    if (sum(c("BP", "CHR") %in% col_headers) == 2) {
+    #non-bi-allelic SNPs can share same BP position so don't check if selected
+    if (sum(c("BP", "CHR") %in% col_headers) == 2 & bi_allelic_filter) {
         message("Checking for SNPs with duplicated base-pair positions.")
         # Try to remove duplicated Positions
         data.table::setkey(sumstats_dt, BP, CHR)
+        #remove indels from this check if sleected since these can have same pos
+        indel_dt <- data.table::data.table()
+        if(sum(c("A1", "A2") %in% col_headers) == 2 & indels ){
+            #identify Indels based on num char in A1, A2
+            num_indels <- nrow(sumstats_dt[(nchar(A1)>1 | nchar(A2)>1),])
+            if(num_indels>0){
+                msg <- paste0("Found ",
+                              nrow(num_indels), " Indels. These won'",
+                              "t be checked for duplicates based on base-pair ",
+                              "position as there can be multiples.",
+                              "\nWARNING If your sumstat ",
+                              "doesn't contain Indels, set the ",
+                              "indel param to FALSE & rerun ",
+                              "MungeSumstats::format_sumstats()")
+                message(msg)
+                indel_dt <- sumstats_dt[(nchar(A1)>1 | nchar(A2)>1),]
+                #run generic dup check on indels
+                indel_dt_rtn <- check_dup_row(
+                    sumstats_dt = indel_dt,
+                    path = path,
+                    log_folder_ind = log_folder_ind,
+                    check_save_out = check_save_out,
+                    tabix_index = tabix_index,
+                    nThread = nThread,
+                    log_files = log_files
+                )
+                indel_dt <- indel_dt_rtn$sumstats_dt
+                # update values
+                log_files <- indel_dt_rtn$log_files
+                sumstats_dt <- sumstats_dt[!(nchar(A1)>1 | nchar(A2)>1),]
+            }
+        }
         dups <- duplicated(sumstats_dt, by = data.table::key(sumstats_dt))
         if (sum(dups) > 0) {
             msg <- paste0(
@@ -57,12 +92,35 @@ check_dup_bp <- function(sumstats_dt,
             sumstats_dt <- unique(sumstats_dt,
                 by = data.table::key(sumstats_dt)
             )
+            
+            #join back indels
+            sumstats_dt <-
+                data.table::rbindlist(
+                    list(sumstats_dt, indel_dt))
 
             return(list(
                 "sumstats_dt" = sumstats_dt,
                 "log_files" = log_files
             ))
         }
+        #join back indels
+        sumstats_dt <-
+            data.table::rbindlist(
+                list(sumstats_dt, indel_dt))
+    }
+    else{
+        #run generic dup check
+        sumstats_return <- check_dup_row(
+            sumstats_dt = sumstats_dt,
+            path = path,
+            log_folder_ind = log_folder_ind,
+            check_save_out = check_save_out,
+            tabix_index = tabix_index,
+            nThread = nThread,
+            log_files = log_files
+        )
+        sumstats_dt <- sumstats_return$sumstats_dt
+        log_files <- sumstats_return$log_files
     }
     return(list("sumstats_dt" = sumstats_dt, "log_files" = log_files))
 }