+++ R/validate_parameters.R
@@ -28,6 +28,7 @@ validate_parameters <- function(path,
                                 snp_ids_are_rs_ids,
                                 remove_multi_rs_snp,
                                 frq_is_maf,
+                                indels,
                                 write_vcf,
                                 return_format,
                                 ldsc_format,
@@ -179,6 +180,9 @@ validate_parameters <- function(path,
     }
     if (!is.logical(frq_is_maf)) {
         stop("frq_is_maf must be either TRUE or FALSE")
+    }
+    if (!is.logical(indels)) {
+        stop("indels must be either TRUE or FALSE")
     }
     if (!is.logical(write_vcf)) {
         stop("write_vcf must be either TRUE or FALSE")